fixed search and gclcl
Former-commit-id: 4a5d9d5939ce9066f71cacc377f00fbc457f7a18

+++ app/Http/Controllers/GlobalChargesLclController.php
@@ -268,8 +268,10 @@ class GlobalChargesLclController extends Controller
 
     public function destroyGlobalCharges($id)
     {
-        $global = GlobalChargeLcl::find($id);
-        $global->delete();
+        $global = GlobalChargelcl::find($id);
+        if(!empty($global)){
+            $global->delete();
+        }
     }
 
     public function destroyArr(Request $request)
